+++ src/logan/pickpocket/command/PickPocketCommand.java
@@ -8,13 +8,6 @@ import java.util.List;
 /**
  * Created by Tre on 12/15/2015.
  */
-public abstract class PickPocketCommand {
-
-    public boolean execute(Player player, List<Profile> profiles, Object... args) {
-        return true;
-    }
-
-    public boolean execute(Player player, List<Profile> profiles) {
-        return true;
-    }
+public interface PickpocketCommand {
+    void execute(Player player, List<Profile> profiles, Object... args);
 }